Source To Target Mapping Vocabulary
We propose an OWL-based mapping vocabulary S2TMAP illustrated in this figure. A mapping between a source and target TBoxes is represented as an instance of the class map:MapDataset. The source and target TBoxes are dened by initializing map:TBox and those TBoxes are connected to the mapping dataset using the properties map:sourceTBox and map:targetTBox respectively. A concept mapping (an instance of map:ConceptM apper) is used to map between a source and a target concepts (instances of map:Concept). A concept mapping is connected to a mapping dataset using map:mapDataset property. The instances of the class map:Concept are all concepts (dened in Equation 1) of source and target TBoxes. The relationship between the concepts can be either rdfs:subClassOf, or rdfs:subPropertyOf, or owl:equivalentClass, and it is connected to the concept mapping via map:rela tion property. The class
setl:Operation includes all the Execution Layer operations described in Section 5.2. The operation to implement the concept mapping in ABox level is dened through map:operation. The common property between the source and target concepts are connected to the concept mapping using map:commonSourceProperty and map:commonTargetProperty properties. The instances of the classmap:Property are all properties (dened in Equation 2) of source and target TBoxes. The type of target instance IRIs is stated using the property map:Target InstanceIriUniqueValueType. If the type is either map:Property or map:Expression, then the property or expression to be used to generate the IRIs is given by map:targetInstanceIriUniquevalue. To map in property level, a property mapping (an instance of map:PropertyMapping) is used. The association between a property mapping and a concept mapping is dened by map:conceptMapping. The target property of the property mapping is stated using map:targetProperty and that target property can either be mapped with a source property or an expression. The source-type of target property is determined through map:source4TargetPropertyValue
property, and the value is dened bymap:source4TargetPropertyValue.